social_network_backend/pyproject.toml
[tool.poetry]
name = "code"
version = "0.1.0"
description = "Social Network Web Application"
authors = ["xenn <rishi734438@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4"
Django = "5.0.7"
django-cors-headers = "3.7.0"
django-environ = "0.11.2"
django-filter = "23.2"
django-storages = "1.13.2"
djangorestframework = "3.15.1"
djangorestframework-simplejwt = "5.2.2"
Pillow = "10.2.0"
pre-commit = "3.3.3"
psycopg2 = "2.9.7"
urllib3 = "^2.2.2"
chardet = "^5.2.0"
idna = "^3.7"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
